교수 교적 조회 화면

+++ src/main/java/com/msa4lms/domain/auth/services/AuthService.java
@@ -61,7 +61,7 @@ public class AuthService {
 
         String extractRefreshToken = refreshTokenOptional.get();
 
-        int id = Integer.parseInt(jwtProvider.extractClaims(extractRefreshToken).getSubject());
+        long id = Long.parseLong(jwtProvider.extractClaims(extractRefreshToken).getSubject());
 
         User user = userMapper.findByPk(id);
 
@@ -131,7 +131,7 @@ public class AuthService {
 
 
     // 비밀번호 변경
-    public void changePassword(int id, PasswordChangeReq req){
+    public void changePassword(long id, PasswordChangeReq req){
 
         User user = userMapper.findByPk(id);
